Update FileStream seek, tell and size to support files greater than 4GB, and max size is 2^63.

+++ cocos/platform/CCPosixFileStream.h
@@ -109,11 +109,11 @@ public:
     bool open(const std::string& path, FileStream::Mode mode) override;
     int close() override;
 
-    int seek(long offset, int origin) override;
+    long seek(int64_t offset, int origin) override;
     int read(void* buf, unsigned int size) override;
     int write(const void* buf, unsigned int size) override;
-    int tell() override;
-    long long size() override;
+    int64_t tell() override;
+    int64_t size() override;
     bool isOpen() const override;
 
 private:
